Replace query strings by query objects in DataStore
This commits does the following:
- it creates Query objects to represent queries. The DML versions of
  those objects (INSERT, UPDATE and DELETE queries) can be inspected
  to extract the actual values modified.
- the existing query builder (returned by `DataStore#query`) is
  retrofitted to produce those new objects (which involves lots of
  change to the implementation, but relatively few to the API).
- `DataStore` is modified to exclusively use those query objects
  and not query strings anymore. All users of DataStore are updated
  accordingly.

The goal of this change is to allow for a mechanism like CDC to
intercept DataStore writes and extract the actual data written.
This patch does not add hooks into DataStore for CDC to do this,
but such hooks should now be easily to add. CDC would then be
able to handle the `BoundInsert`, `BoundUpdate` and `BoundDelete`
queries.

Note that as of this patch, the CQL API still uses query strings
as it uses the `Persistence` interface directly, not `DataStore`.
The longer term plan is to push the new query objects to the
`Persistence` API too, but doing so require query parsing and
to handle a lot more of the diversity of features that CQL
exposes. This patch should be considered a first step in that
direction.